requirements.txt
# Core
django==4.1.12
python-dotenv==1.0.0
djangorestframework==3.14.0

# UI
django-admin-black==1.0.8

# Deployment
whitenoise==6.5.0
gunicorn==21.2.0

# DB Layer
# psycopg2-binary
# mysqlclient
